% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPolytomyWithMaxCR.R
\name{fun.AdjustPolytomy}
\alias{fun.AdjustPolytomy}
\title{Adjust polytomy with maximum coalescent rate}
\usage{
fun.AdjustPolytomy(RRResult)
}
\arguments{
\item{RRResult}{The list of result obtained from Np_Estimator.}
}
\value{
A dataframe containing lower and upper boundary.
}
\description{
Adjust polytomy with maximum coalescent rate
}
\examples{
library(TarCA)
load(system.file("Exemplar","Exemplar_TCA.RData",package = "TarCA"))
fun.AdjustPolytomy(ExemplarData_1)

}
