% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub__CheckInputFile.R
\name{fun.CheckInput}
\alias{fun.CheckInput}
\title{Check input file}
\usage{
fun.CheckInput(TTTree, AAAnn, FFFileout)
}
\arguments{
\item{TTTree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{AAAnn}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}

\item{FFFileout}{A path to output (e.g. tmp.RData).}
}
\value{
A processed tree.
}
\description{
Check input file
}
