% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub__GetPhyloWithPairMRCA.r
\name{fun.GetPhyloWithPairMRCA}
\alias{fun.GetPhyloWithPairMRCA}
\title{Recover a phylogeny with pairwise MRCA of internal nodes}
\usage{
fun.GetPhyloWithPairMRCA(DDData)
}
\arguments{
\item{DDData}{A dataframe containing pairs of tips and corresponding MRCA.}
}
\value{
A tree file of class "phylo"
}
\description{
Recover a phylogeny with pairwise MRCA of internal nodes
}
\examples{
library(dplyr)
library(tidyr)
tmp.tree <- ape::rtree(10)
tmp1 <- sample(1:tmp.tree$Nnode,5)
tmp2 <- data.frame(T1=gl(5,5),T2=gl(5,1,25)) \%>\%
filter(T1!=T2) \%>\%
mutate(T1=tmp1[T1],T2=tmp1[T2]) \%>\%
mutate(MRCA=castor::get_pairwise_mrcas(tmp.tree,T1,T2))
tmp2 \%>\% fun.GetPhyloWithPairMRCA

}
