% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeShortEdge.R
\name{fun.MergePolytomy}
\alias{fun.MergePolytomy}
\title{Merge short edge into polytomy}
\usage{
fun.MergePolytomy(TTTree, EEEpsilon = 1e-05)
}
\arguments{
\item{TTTree}{A tree file of class "phylo".}

\item{EEEpsilon}{Non-negative numeric, specifying the maximum edge length for an edge to be considered ``short'' and thus to be eliminated. Typically 0 or some small positive number.}
}
\value{
A list containing processed tree and distribution of the edge length.
}
\description{
Merge short edge into polytomy
}
