% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeShortEdge.R
\name{fun.PlotEdgeLengthDistribution}
\alias{fun.PlotEdgeLengthDistribution}
\title{Plot the distribution of edge length}
\usage{
fun.PlotEdgeLengthDistribution(TTTree)
}
\arguments{
\item{TTTree}{A tree file of class "phylo".}
}
\description{
Plot the distribution of edge length
}
