% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveHugePolytomy.R
\name{fun.RemoveHugePolytomies}
\alias{fun.RemoveHugePolytomies}
\title{Remove huge polytomies from the phylogeny}
\usage{
fun.RemoveHugePolytomies(TTTree, NNNum)
}
\arguments{
\item{TTTree}{A tree file of class "phylo".}

\item{NNNum}{The maximum size of polytomy node.}
}
\value{
The processed tree.
}
\description{
Remove confusing tips from huge polytomy.
}
