% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Np_Estimator.R
\name{Np_Estimator}
\alias{Np_Estimator}
\title{Estimate Np using cell phylogeny for each sub-population.}
\usage{
Np_Estimator(
  PathToData = NULL,
  Tree = NULL,
  Ann = NULL,
  Fileout = NULL,
  ReturnNp = TRUE
)
}
\arguments{
\item{PathToData}{A RData file including tmp.tree, tmp.ann and tmp.fileout.}

\item{Tree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{Ann}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}

\item{Fileout}{A path to output (e.g. tmp.RData).}

\item{ReturnNp}{Return the result or not (TRUE by default).}
}
\value{
A list containing Tree, Ann, PureNode, PureNode2Organ and EffN. The estimated Np are store in EffN.
}
\description{
Estimate Np using cell phylogeny for each sub-population.
}
\examples{
library("TarCA")
load(system.file("Exemplar","Exemplar_TCA.RData",package = "TarCA"))
tmp.res <- Np_Estimator(
Tree = ExemplarData_1$Tree,
Ann = ExemplarData_1$Ann,
Fileout = NULL,
ReturnNp = TRUE
)
}
